Purchases optimization -- other bug fixes

+++ database/migrations/2023_01_15_224712_create_purchases_table.php
@@ -25,8 +25,8 @@ return new class extends Migration
             $table->double('qte');
             $table->double('unit_price');
             $table->double('buying_price');
-            $table->decimal('tax_amount');
-            $table->decimal('grand_total');
+            $table->double('tax_amount');
+            $table->double('grand_total');
             $table->tinyInteger('status')->default('0')->comment('0=Pending,1=Approved');
             $table->integer('created_by')->nullable();
             $table->integer('updated_by')->nullable();
